Implement view patterns.
Also simplify pattern match desugaring.
Use view patterns for matching literals.

+++ tests/View.hs
@@ -2,11 +2,10 @@ module View(main) where
 import Prelude
 
 f :: Int -> Int
-f ((> (10::Int)) -> True) = 1
+f ((> 10) -> True) = 1
 f _ = 0
 
 main :: IO ()
 main = do
   print (f 20)
   print (f 10)
-
